Split enum test. Rename dot.js to template.js
