+++ lib/index.spec.js
@@ -507,7 +507,32 @@ describe('request error handler', function () {
           ]))
       })
 
-      it('enum', function () {
+      it('enum (one value)', function () {
+        return test(
+          {
+            'application/json': {
+              schema: JSON.stringify({
+                type: 'string',
+                enum: ['a']
+              })
+            }
+          },
+          '"b"',
+          JSON_HEADERS
+        )
+          .then(validateJsonResponseErrors([
+            {
+              data: 'b',
+              dataPath: '',
+              keyword: 'enum',
+              message: 'No enum match for: a',
+              schema: ['a'],
+              type: 'json'
+            }
+          ]))
+      })
+
+      it('enum (mutiple values)', function () {
         return test(
           {
             'application/json': {
