+++ lib/messages/form.js
@@ -1,4 +1,4 @@
-var compile = require('../dot').compile
+var compile = require('../template').compile
 var json = require('./json')
 
 module.exports = {
