+++ lib/messages/json.js
@@ -1,4 +1,4 @@
-var compile = require('../dot').compile
+var compile = require('../template').compile
 
 // Originally based on: https://github.com/geraintluff/tv4/tree/master/lang
 module.exports = {
